test(event-rules): split dispatch tests to meet LOC limits
Issue: #2679

Split test_event_rules.py (578 lines) to meet LOC exception limit (480 lines):
- Move DispatchIntent publishing tests to test_event_rules_dispatch.py
- Keep core event rules tests in test_event_rules.py
- Final counts: test_event_rules.py (417 lines), test_event_rules_dispatch.py (172 lines)

Verification:
- All 36 tests pass
- LOC check passes with ENFORCE_LOC_LIMITS=true
